+++ CHANGELOG.md
@@ -110,6 +110,7 @@ and be accessible according to the permissions that were configured for the role
 - console: support tracking VOLATILE functions as mutations or queries (close #6228)
 - console: show only compatible postgres functions in computed fields section (close #5155) (#5978)
 - console: added export data option on browse rows page (close #1438 #5158)
+- console: add session argument field for computed fields (close #5154) (#5610)
 - cli: add missing global flags for seed command (#5565)
 - cli: allow seeds as alias for seed command (#5693)
 - cli: fix action timeouts not being picked up in metadata operations (#6220)